requirements.txt
streamlit
google-generativeai
PTL
python-dotenv